inboxwhisper/requirements.txt
# Core
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# LangGraph & LangSmith
langgraph>=0.1.0
langsmith>=0.1.0

# Microsoft SSO / Graph
msal>=1.20.0
msal-extensions>=1.0.0

# Notebook / Dev
jupyterlab
notebook
ipykernel

# Optional (UI)
streamlit

# Utilities
pandas
python-dateutil

playwright==1.44.0
beautifulsoup4>=4.9.0
requests>=2.28.0
